Update the module to handle multiple schema connections
