+++ README.md
@@ -11,14 +11,18 @@ library. This is achieved saving the document attributes as a JSON string on the
 ```javascript
 const createDb = require('@miroculus/anaconda-io-db')
 
-const db = createDb({
-  filepath: DB_LOCATION, // DB_LOCATION can be any of the values defined here:
-                         // https://github.com/mapbox/node-sqlite3/wiki/API#new-sqlite3databasefilename-mode-callback
-  tableName: 'protocols', // REQUIRED: Table were to save the documents
-  primaryKey: 'id', // OPTIONAL, defaults to 'id'; is the column name of the documents primary key
-  indexes: ['active'] // OPTIONAL, you can add additional keys to be indexed
-                      // This will allow you to find documents by the indexed column
-})
+// The DB location can be any of the values defined here:
+// https://github.com/mapbox/node-sqlite3/wiki/API#new-sqlite3databasefilename-mode-callback
+const location = DB_LOCATION
+
+const db = createDb(location, [
+  {
+    tableName: 'protocols', // REQUIRED: Table were to save the documents
+    primaryKey: 'id', // OPTIONAL, defaults to 'id'; is the column name of the documents primary key
+    indexes: ['active'] // OPTIONAL, you can add additional keys to be indexed
+                        // This will allow you to find documents by the indexed column
+  }
+])
 
 // Be sure to let the DB successfully connect before trying to use it
 db.connect().then(() => {
@@ -29,7 +33,7 @@ db.connect().then(() => {
 ### Create a document
 
 ```javascript
-await db.create({
+await db.protocols.create({
   id: 123123123,
   name: 'Some Protocol',
   active: true,
@@ -48,11 +52,11 @@ as is.
 ### Find a single document
 
 You can find a document by id following the next example. This result will be
-the data that you passed to `db.create(...)`.
+the data that you passed to `db.protocols.create(...)`.
 
 ```javascript
-const result = await db.findOne(123123123)
-const result = await db.findOne({ active: true }) // Using an indexed column
+const result = await db.protocols.findOne(123123123)
+const result = await db.protocols.findOne({ active: true }) // Using an indexed column
 ```
 
 ### Delete documents
@@ -62,8 +66,8 @@ always returns `undefined`, without taking into account if actually deleted
 something.
 
 ```javascript
-await db.destroy(123123123) // This will delete the document with id `123123123`
-await db.destroy({ active: true }) // This will delete ALL the documents that have `active: true`
+await db.protocols.destroy(123123123) // This will delete the document with id `123123123`
+await db.protocols.destroy({ active: true }) // This will delete ALL the documents that have `active: true`
 ```
 
 ### Update one or multiple documents
@@ -72,13 +76,13 @@ To update a document by ID follow the next example, and this function will retur
 new update document.
 
 ```javascript
-await db.update(123123123, { newValue: 123 })
+await db.protocols.update(123123123, { newValue: 123 })
 ```
 
 To delete an attribute on the given document, give it a value of `undefined`:
 
 ```javascript
-await db.update(123123123, { newValue: undefined })
+await db.protocols.update(123123123, { newValue: undefined })
 ```
 
 You can also update multiple documents in the same operation. To do that, you can
@@ -87,7 +91,7 @@ and set them all the same value. The following example will change all the docum
 that have `active: true` and set them `newValue: 1234`:
 
 ```javascript
-await db.update({ active: true }, { newValue: 1234 })
+await db.protocols.update({ active: true }, { newValue: 1234 })
 ```
 
 ### List Documents
@@ -96,8 +100,8 @@ The following method will return an array with all the documents saved on the
 DB.
 
 ```javascript
-await db.find()
-await db.find({ active: true }) // only find documents with `active: true`
+await db.protocols.find()
+await db.protocols.find({ active: true }) // only find documents with `active: true`
 ```
 
 ### Destroy all documents
@@ -106,7 +110,7 @@ If you want, you can delete absolutely all documents saved on the DB using
 the following command:
 
 ```javascript
-await db.empty()
+await db.protocols.empty()
 ```
 
 ### Disconnection
